line reporter adjustment: extract for reuse
